requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain-community
langchain-openai
requests
beautifulsoup4
chromadb
tiktoken